added package & minor fixes
