+++ .wpilib/wpilib_preferences.json
@@ -1,0 +1,6 @@
+{
+    "enableCppIntellisense": false,
+    "currentLanguage": "java",
+    "projectYear": "2019",
+    "teamNumber": 5968
+}
