--- DriveMode.java
@@ -1,6 +1,0 @@
-public enum DriveMode {
-    DRIVERCONTROL,
-    AUTODRIVINGTRAIGHT,
-    AUTOROTATING,
-    LINEALIGNMENT
-}
