--- IEncoder.java
@@ -1,11 +1,0 @@
-public interface IEncoder {
-
-    public void setInverted(boolean inverted);
-
-    public void setDistancePerPulse(double distance);
-
-    public double getDistance();
-    
-    public void reset();
-
-}
