--- IGyroscope.java
@@ -1,9 +1,0 @@
-public interface IGyroscope {
-    
-    public double getPitch();
-
-    public double getRoll();
-
-    public double getYaw();
-    
-}
