--- IHook.java
@@ -1,9 +1,0 @@
-public interface IHook {
-
-    public void grabPanel();
-
-    public void releasePanel();
-
-    public void toggleGrabbing();
-    
-}
