--- ILauncher.java
@@ -1,9 +1,0 @@
-public interface ILauncher {
-
-    public void pullInCargo();
-
-    public void releaseCargo();
-
-    public void periodic();
-
-}
