--- IRobotMode.java
@@ -1,7 +1,0 @@
-public interface IRobotMode {
-
-    public void init();
-
-    public void periodic();
-
-}
