+++ src/main/java/org/usfirst/frc/team5968/robot/DriveMode.java
@@ -1,0 +1,6 @@
+public enum DriveMode {
+    DRIVERCONTROL,
+    AUTODRIVINGTRAIGHT,
+    AUTOROTATING,
+    LINEALIGNMENT
+}
