+++ src/main/java/org/usfirst/frc/team5968/robot/IEncoder.java
@@ -1,0 +1,11 @@
+public interface IEncoder {
+
+    public void setInverted(boolean inverted);
+
+    public void setDistancePerPulse(double distance);
+
+    public double getDistance();
+    
+    public void reset();
+
+}
