+++ src/main/java/org/usfirst/frc/team5968/robot/IGyroscope.java
@@ -1,0 +1,9 @@
+public interface IGyroscope {
+    
+    public double getPitch();
+
+    public double getRoll();
+
+    public double getYaw();
+    
+}
