+++ src/main/java/org/usfirst/frc/team5968/robot/IHook.java
@@ -1,0 +1,9 @@
+public interface IHook {
+
+    public void grabPanel();
+
+    public void releasePanel();
+
+    public void toggleGrabbing();
+    
+}
