+++ src/main/java/org/usfirst/frc/team5968/robot/ILauncher.java
@@ -1,0 +1,9 @@
+public interface ILauncher {
+
+    public void pullInCargo();
+
+    public void releaseCargo();
+
+    public void periodic();
+
+}
