+++ src/main/java/org/usfirst/frc/team5968/robot/IRobotMode.java
@@ -1,0 +1,7 @@
+public interface IRobotMode {
+
+    public void init();
+
+    public void periodic();
+
+}
